% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_wrapper.R
\name{unit_r}
\alias{unit_r}
\title{Create Unit tests for R Files using bug_ai}
\usage{
unit_r(input_file_or_dir, output_dir)
}
\arguments{
\item{input_file_or_dir}{path to input_file or dir}

\item{output_dir}{Directory to save the documented R file.}
}
\value{
Path to the documented R file.
}
\description{
This function documents an R file using the \code{bug_ai} Python tool.
}
